Add layers to calling on controller actions to insure they exist
Need some additional steps to catch when a route or controller action
isn't found so the caller can take the appropriate actions.  Likely,
call on a 404 page to be displayed.

This is a major API change, as the caller now has to call run(), then
request a status, then finally execute() if all is good to go.